ai_structuring_backend/requirements.txt
# S4Carlisle Pre-Editor v3 Backend Requirements

# Web Framework
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Document Processing
python-docx>=1.1.0

# AI Classification
google-genai>=0.3.0

# Task Queue (Production)
celery>=5.3.0
redis>=5.0.0
flower
# Data Validation
dataclasses-json>=0.6.0

# Utilities
python-dotenv>=1.0.0

# Development & Testing
pytest>=8.0.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0
psycopg2-binary>=2.9

# Production Server
gunicorn>=21.0.0
